Move nan_percentile to be a general utility function
